requirements.txt
streamlit
Pillow
openpyxl